hot fix

+++ 0x0D-preprocessor/3-function_like_macro.h
@@ -1,6 +1,6 @@
 #ifndef FUNC_H
 #define FUNC_H
 
-#define ABS(x) (((x) < 0) ? (-(double)(x)) : ((double)(x)))
+#define ABS(x) (((x) < 0) ? (-(x)) : (x))
 
 #endif /* FUNC_H */
